links-metrics-api/requirements.txt
-f https://s3pypi.mettel-automation.net/igzpackages/index.html
igzpackages==2.0.1
pymongo==3.11.4
Quart-CORS==0.1.3
quart-openapi==1.4.3
pyhumps==1.3.1